Split the code in controller.go
